src/main/java/com/beta/replyservice/commands/EncodeMd5Command.java
package com.beta.replyservice.commands;

import java.math.BigInteger;
import java.security.MessageDigest;
import java.security.NoSuchAlgorithmException;

public class EncodeMd5Command implements Command {
    @Override
    public String execute(String inputStr) {
        return getMD5(inputStr);
    }

    private String getMD5(String input) {
        if (input == null) {
            return input;
        }
        try {
            MessageDigest md = MessageDigest.getInstance("MD5");
            byte[] messageDigest = md.digest(input.getBytes());
            BigInteger no = new BigInteger(1, messageDigest);
            String hashtext = no.toString(16);
            while (hashtext.length() < 32) {
                hashtext = "0" + hashtext;
            }
            return hashtext;
        } catch (NoSuchAlgorithmException e) {
            throw new RuntimeException(e);
        }
    }
}
